reset play time if user quits game and starts up the same game
There was a "bug" that would make it so the timestamp would retain the same time from before if the user quit the game and then started it back up. This could be kept as a feature (e.g. the game crashed) but I personally would rather make it reset.